Уточнены коэффициенты по настройкам двигателя

+++ DrillingRig.ConfigApp/AppControl/EngineSettingsSpace/EngineSettingsPartWritable.cs
@@ -3,9 +3,9 @@
 		public ushort? Inom { get; set; }
 		public ushort? Nnom { get; set; }
 		public ushort? Nmax { get; set; }
-		public uint? Pnom { get; set; }
-		public ushort? CosFi { get; set; }
-		public ushort? Eff { get; set; }
+		public double? Pnom { get; set; }
+		public double? CosFi { get; set; }
+		public double? Eff { get; set; }
 		public ushort? Mass { get; set; }
 		public ushort? MmM { get; set; }
 		public ushort? Height { get; set; }
